+++ src/lib.rs
@@ -426,6 +426,9 @@
 
 #![doc(html_no_source)]
 
+#[cfg(not(target_family = "unix"))]
+compile_error!("Sorry, this project currently only supports the Unix platform.");
+
 mod algorithm_id;
 mod callback;
 mod context;